Refactor View to Display per David comments
