+++ src/main/java/com/ccs/baby/debug/BabyAsmSyntaxHighlighter.java
@@ -1,4 +1,4 @@
-package com.ccs.baby.ui;
+package com.ccs.baby.debug;
 
 import java.awt.Color;
 import java.util.regex.Matcher;
