+++ src/main/java/com/ccs/baby/ui/display/DisplayControl.java
@@ -1,14 +1,19 @@
-package com.ccs.baby.ui;
+package com.ccs.baby.ui.display;
+
+import com.ccs.baby.ui.CrtPanel;
+import com.ccs.baby.ui.SwitchPanel;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
-// change to display the Control
-public class ViewControl implements ActionListener {
+/**
+ * This class is responsible for displaying the control.
+ */
+public class DisplayControl implements ActionListener {
     CrtPanel crtPanel;
     SwitchPanel switchPanel;
 
-    public ViewControl(CrtPanel aCrtPanel, SwitchPanel aSwitchPanel) {
+    public DisplayControl(CrtPanel aCrtPanel, SwitchPanel aSwitchPanel) {
         crtPanel = aCrtPanel;
         switchPanel = aSwitchPanel;
     }
